requirements.txt
# Data pipeline
requests>=2.31.0
ijson>=3.2.3
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
anthropic>=0.25.0
boto3>=1.34.0
tqdm>=4.66.0

# API
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-jose[cryptography]>=3.3.0
pydantic>=2.7.0

# Testing
pytest>=8.2.0
httpx>=0.27.0   # for FastAPI test client
